+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 ### New features
 
 * Pdf summary now shows summary for the report duration
+* Now you can programmatically open new transaction page with prefilled fields.
+  * Example for creating $300 transaction in an USD account:
+
+    `flow-mn:/transaction/new?amount=300&fromAccountUuid=2627442e-b741-4ab7-93a8-d82178007519`
+  * All fields are passed by uri query params
+  * To retrieve the account id, use json export
+  * For valid fields, see [schema](./schemas/20250919-transaction-programmable-object.json)
 
 ## 0.15.0